Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -66,7 +66,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return view ('questions.edit', compact('question')); // compact('question') is the same as ['question' => $question]
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
+        return view ('questions.edit', compact('question'));
+
     }
 
     /**
@@ -91,8 +95,11 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
-        $question->delete();
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, "Access denied");
+        }
 
+        $question->delete();
         return redirect()->route('questions.index')->with('success', "Your question has been deleted.");
     }
 }
